+++ Game/Level 2/JS.js
@@ -22,7 +22,8 @@ async function audio() {
         //Defines the audio for the typing sound, and audio.
         var sound = {
             music: new Howl({
-                src: ['../../Media/Music/Black Fingerprint - Dylan Owen.wav'],
+                //CHANGE IT
+                src: ['../../Media/Music/Mystery - Wee Free Music.wav'],
                 //src: ['../../Media/Sound/Wii crashing sound.mp3'],
                 html5: true,
                 volume: 0,
@@ -33,7 +34,7 @@ async function audio() {
         };
 
         var music = sound.music.play();
-        sound.music.fade(0,0.5,10000,music);
+        sound.music.fade(0,0.5,5000,music);
         var typing_sound = sound.typing_sound.play()
     }
 }
@@ -66,7 +67,7 @@ async function typing() {
     //console.log("Debug function")
     delay_text = "100"
     await sleep(5000);
-    proceed_to_typer_element("Computer: We need to find a way to the mainframe.");
+    proceed_to_typer_element("Computer: I'm out of ideas.");
     //var typing_sound = sound.typing_sound.play()
     await sleep(2000);
     //sound.stop()
@@ -74,18 +75,7 @@ async function typing() {
     delay_text = '65'
     color_text = "white"
     await sleep(5000);
-    proceed_to_typer_element("Puzzle: The mainframe is on a circuit that leads to no other component. Click your answer.");
-    await sleep(7500);
-    document.getElementById("puzzle_image").src="../../Media/Images/Puzzle 1/PCB.png";
-    await sleep(1000);
-    document.getElementById("chip_4").src='../../Media/Images/Puzzle 1/Chip.png';
-    await sleep(500);
-    document.getElementById("chip_2").src='../../Media/Images/Puzzle 1/Chip.png';
-    await sleep(500);
-    document.getElementById("chip_3").src='../../Media/Images/Puzzle 1/Chip.png';
-    await sleep(500);
-    document.getElementById("chip_1").src='../../Media/Images/Puzzle 1/Chip.png';
-    await sleep(500);
-    document.getElementById("chip_5").src='../../Media/Images/Puzzle 1/Chip.png';
+    proceed_to_typer_element("Puzzle: I'm out of ideas");
+    await sleep(30000);
 }
 typing()