+++ Game/Level 3/JS.js
@@ -22,7 +22,7 @@ async function audio() {
         //Defines the audio for the typing sound, and audio.
         var sound = {
             music: new Howl({
-                src: ['../../Media/Music/Black Fingerprint - Dylan Owen.wav'],
+                src: ['../../Media/Music/Mystery - Wee Free Music.wav'],
                 //src: ['../../Media/Sound/Wii crashing sound.mp3'],
                 html5: true,
                 volume: 0,
@@ -33,7 +33,7 @@ async function audio() {
         };
 
         var music = sound.music.play();
-        sound.music.fade(0,0.5,10000,music);
+        sound.music.fade(0,0.5,5000,music);
         var typing_sound = sound.typing_sound.play()
     }
 }
@@ -66,7 +66,7 @@ async function typing() {
     //console.log("Debug function")
     delay_text = "100"
     await sleep(5000);
-    proceed_to_typer_element("Computer: We need to find a way to the mainframe.");
+    proceed_to_typer_element("Computer: Hold on. There's a password with a puzzle on it.");
     //var typing_sound = sound.typing_sound.play()
     await sleep(2000);
     //sound.stop()
@@ -74,18 +74,24 @@ async function typing() {
     delay_text = '65'
     color_text = "white"
     await sleep(5000);
-    proceed_to_typer_element("Puzzle: The mainframe is on a circuit that leads to no other component. Click your answer.");
-    await sleep(7500);
-    document.getElementById("puzzle_image").src="../../Media/Images/Puzzle 1/PCB.png";
+    proceed_to_typer_element("Puzzle: Fifteen people are trapped aboard a ship that's going to sink in exactly 20 minutes. Their only chance for survival is the five-person life raft stowed on their vessel. To make matters worse the waters around the ship are teeming with man-eating sharks. So swimming to safety is out of the question. A round-trip to the nearest island and back to the ship takes nine minutes on the raft. How many people will live to see dry land?");
+    await sleep(30000);
+    //var answer = prompt("Your answer.")
+    let answer_input = document.getElementById("answer");
+    answer_input.classList.toggle("show");
     await sleep(1000);
-    document.getElementById("chip_4").src='../../Media/Images/Puzzle 1/Chip.png';
-    await sleep(500);
-    document.getElementById("chip_2").src='../../Media/Images/Puzzle 1/Chip.png';
-    await sleep(500);
-    document.getElementById("chip_3").src='../../Media/Images/Puzzle 1/Chip.png';
-    await sleep(500);
-    document.getElementById("chip_1").src='../../Media/Images/Puzzle 1/Chip.png';
-    await sleep(500);
-    document.getElementById("chip_5").src='../../Media/Images/Puzzle 1/Chip.png';
+    button_answer.classList.toggle("show");
 }
 typing()
+
+function check_answer() {
+    let answer = document.getElementById("answer").value
+    if ((answer === "2") || (answer === "Two") || (answer === "two")) {
+        //console.info("Correct answer!");
+    }
+
+    else {
+        //console.info("Wrong answer!");
+        window.open('../GameOver/HTML.html','_self')
+    }
+}